+++ src/onegov.agency/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Restyles the hidden contents views.
+  [msom]
+
 1.1.0 (2018-11-26)
 ~~~~~~~~~~~~~~~~~~~~